Flatten columns, column renaming
The multi-index columns have been clumsy for some folks. Now return a
dataframe with flat columns and a string prefix indicating the column
group.

It is still possible to convert to a hierarchical index, select a
particular group, or drop the group level in post-processing. This usage
is shown in the example.

Also rename sidecar -> json following the suggestion in (#15).