Flatten columns, column renaming (#16)
The multi-index columns have been clumsy for some folks. Now return a dataframe with flat columns and a string prefix indicating the column group.

It is still possible to convert to a hierarchical index, select a particular group, or drop the group level in post-processing. This usage is shown in the example.

Other changes:
- rename sidecar -> json following the suggestion in (#14).
- by default, search for inherited metadata until a `dataset_description.json` is found, following the suggestion in (#15).
- Fix `incremental` mode by specifying correct path and modified time column names.