Downgrade minimal required python to 3.11
Requiring python>=3.12 is a bit of a burden. It's only required for
`find_bids_datasets`, which uses `Path.walk`. Instead, downgrade minimum
python to 3.11 and add a guard on this function in case python<3.12.

Note, we could look into downgrading further. The next block would be
`get_column_names` returns a `StrEnum`, which was introduced in 3.11.
I'm hesitant to remove this though, because being able to treat these
enum fields as native strings is nice.